Allow for dynammic cancellation of requests (#62)
If client goes away we should also cancel any
underlying requests going on. This allows for
users to specify a context.Context in the client
configuration which will cancel any underlying
requests in case the context is cancelled. If the
context is grabbed from the server-side
*http.Request object, this means we will cancel any
operations being performed when the client cancels
the request.