Use BaseURL instead of URL
We don't want to set path or query on the base client,
so use BaseURL instead to grab just scheme and host.

We believe this won't break any usage but we should
pay attention when updating to this go-clients version
to ensure no one was trying to set the path on the
base client.